move rpc methods to room.
